pyproject.toml
[tool.poetry]
name = "defaultbottemplate"
version = "0.1.0"
description = ""
authors = ["arthu <arthur_air@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
loguru = "^0.7.0"
sqlalchemy = "^2.0.15"
asyncpg = "^0.27.0"
pyyaml = "^6.0"
pydantic = "^1.10.8"
openpyxl = "3.0.10"
django = "^4.2.1"
environs = "^9.5.0"
aiogram = "3.0b7"
redis = "^4.5.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
